feat: Initial support for module federation
This allows us to import the OpenshiftStreams react component.

* Move port and module name to `package.json`
* Rename the app
* provide barrel export for OpenshiftStreams
* introduce the bootstrap indirection
* add module federation config to webpack.common.js
* explicitly set the `publicPath` in web pack (required by module
  federation)

Note that non-dev configuration is broken - this needs fixing in a 
later PR
